cli: support get-job and wait-job
get-job simply prints out all of the information received from the
pinpoing server, and may need further work in the future if the response
proto ends up having very verbose information in it.

wait-job polls the backend until the job is no longer PENDING or
RUNNING, printing out information similar to get-job in the meantime.
The -quiet flag prevents the information from being printed out, since
it may be too noisy for some usecases.

In the future the logic for wait-job may be integrated into other parts
of the CLI, e.g. via a `-wait` option for commands which submit jobs.
For now, users have to manually call wait-job separately.

Change-Id: I3275e13d0f75f9f0e8d80d872ce1da2410d824c4