caching-backend: conf-creator: wait until the node is configured in UFS
As the init container of caching backends, right now conf-creator
creates some non-operational conf files for Nginx and keepalived when
the node is not configured in UFS yet. After the UFS configuration done,
we have to restart the POD.

With this patch, it will not exit until the node is configured.

BUG=None
TEST=Built and ran locally with fake 'NODE_IP' and 'NODE_NAME' to
    verify the behavior is as expected, i.e.

Change-Id: I51568b68096b96b4fe5dbec5117fd7a4bf4fc4fa