Fix usage of libtool to work properly on armv6l docker container.
Currently, libtool detects absolute paths for 'sed' and 'grep' at
build time, and expects them to be in the same locations when libtool
is run later. These utilities are not in the same paths across all of
our docker images.

For sed, fix this by depending on the 3pp sed package. Unfortunately
there's no support for transitive tool dependencies, so update all
packages that depend on libtool to also depend on sed.

For grep, we don't currently have a grep 3pp package and this doesn't
seem like a good enough reason to add one. Force libtool to just use
'grep' from $PATH.

Change-Id: Id6167c0487f1b4e89783f1601719299fd008cfc8