Split beautifulsoup4 into two separate wheels.
It supports both Python versions, but via two separate wheels. Prior to
this CL we'd erroneously just download one of the two and upload it as a
py2_py3 wheel.

In this CL, we check the resulting wheel and make sure it supports the
required versions, issuing an error if it doesn't.

Bug: 1212759
Change-Id: I074205de5f4ea76d1179c305da6389a799baad52