[WIC] Fix ADK install process
Bug: b/182061587

Fixes the issue with installing ADK and WinPE on bots. Currently this
cannot support installing the ADK or WinPE to a custom location due to a
bug in the way installer determines if there is another installation on
the system. As a result the module includes cleanup API to clean up the
bots once it is done. However this may not be required as we have new
resources dedicated to the image builder and ideally no other task
would be run on them.

Change-Id: I186caf4a8ee65a0c3c29bb9fa1ec7efe805f900f