upload_debug_symbols: starter boiler plate and gs package fetch
In the initial set up of the project I had not considered that
authentication would be required for the Google storage requests.
This change adds boilerplate code to turn this package into an
application which includes authentication handling. I followed
the same template that our other CIPD packages use.

Aside from the boilderplate, this change includes working functionality
to pull debug symbols from GS. This was tested locally by sending it a
GS path to an uploaded debug.tgz.

BUG=b:195981092,b:197010274
TEST=go test; local execution test

Change-Id: I9696f3dae0a52ed6f06d4756fe67435d8f2a4011