[Weetbix] bq-export: exports the partial rows
We export the rows on a schedule and in the normal cases each row contains a test variant's data in a fixed time range (for example 1 hour). But there are also special cases that a test variant's status changes within that time range.

For example, we run the bq export job every hour at 1AM, 2AM, 3AM and so on. And a test variant's status changes at 1:30AM. When we run the bqexport job at 2AM to export test variants data in the time range [1AM, 2AM):

- If we should only export flaky test variants to the table:
  - If the test variant's status changes from any status to flaky, we should export a row for this test variant with a narrowed time range [1:30, 2AM).
  - If the test variant's status changes from flaky to any other status, we should export a row for this test variant with a narrowed time range [1AM, 1:30AM).

- If we should export all the test variants to the table:
 - we should export 2 rows for this test variants, first one with status before update and time range [1AM, 1:30AM) and second on with status after update and time range [1:30, 2AM)


In rare cases, there could be more than one status updates within the original time range ([1AM, 2AM) in the example), and this CL uses the PreviousStatuses and PreviousStatusUpdateTimes to track them.

Bug: 1263261
Change-Id: I9eb768daebbcc424ec99c6d1f59005ba99bf1a83