[Findit] Do not set Bot-Commit +1 label
Bot-Commit +1 label is only required if no human is reviewing the revert CL. Currently Findit always required human to review CL, so it is not necessary.

Furthermore, setting Bot-Commit+1 results in an error, since Findit does not have the permission to do it.

I think it may be the cause of Findit v2 creating multiple CLs for one culprit (e.g. https://chromium-review.googlesource.com/c/chromium/src/+/3235930/), because the task in the task queue will retry if failed.

Bug:978503,1269907
Change-Id: I1e4fb7111787b385d9f6f7bf4935a68d4efb1860