[recovery]: Using cold_reset, detect if DUT is connected to servo.
This commit is a part of the PARIS project.

For setup that include servo type other than V3, the DUT may be
disconnected from the servo. The servod command cold_reset can be used
to detect such situations. This commit includes a verifier that uses
cold_reset command to detect if DUT is connected.

bug=b:198809511
TEST=build, manual test

Change-Id: Icc8699cdf237129e2d7d0bdf126af11bbd1ca16f