[labpack] Log all inputs in labpack executable.
Labpack currently logs its input arguments on startup.

Change labpack so that the messages that it logs during startup are more
easily distnguished from other similar logs (for example, bbagent log
messages). Also, log all of the inputs to labpack, including the build
state and the positional arguments if there are any.

BUG=b/202156419
TEST=make test

Change-Id: Id2f05c76e20d4a5eb79b0f5b696d0d66fe47b7c5