cache: set proxy_cache_lock_age
http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_cache_lock_age

This a new setting imported since nginx 1.7.8. We don't have it before
we migrate to K8s.

The default value of this setting is 5 seconds, which means if we cannot
complete a downloading in that time, Nginx will pass another request to
upstream (i.e. gs_archive_server). This is not good for us because it's
very often that we have many requests come together for one large file
which need 1~2 minutes to complete. As the result, quite a few requests
are passed to the upstream just for the same file. If all upstream
server are busy, we cannot handle any new requests.

This CL, we change the setting to longer value so only one request can
be passed and all remaining requests will wait patiently.

BUG=b:203360299
TEST=Tested in the caching cluster.

Change-Id: Iee6283ae40b8527f338bcf7ba885a5892778ed10