[recovery]: Verify that value of servod control matches the expected value from config.
This commit is a part of the PARIS project.

The configuration for a servod control specifies the expected value,
as well as value-type of a servod control. This commit verifies that
the actual value returned from servod matches the expected value.

This commit also creates helper functions that can be used to retrieve
values of specific types from servod.

Bug=b:198809511
TEST=build, manual test

Change-Id: If2d44fbd59e96ab587430f860311d9e83a173d71