recovery: add cros_update_device_sku_label action:
This action will determine the correct device_sku label for the DUT
inside the cros_repair (either from DUT itself or inventory). If it
doesn't find the sku label in either places. It will throw an error
indicating the update sku label action fails.

But since this action is just updating the sku label. It is allowed to fail
after recovery.

And this device_sku_label belongs to both cros_repair and cros_deploy, so
this cl updates the default configuration for both the repair and deploy task.

BUG=b:205156117
TEST=manual test

Change-Id: Ia4fd6fd23ffa100c3270a7dd347139d8f30eaa9c