recovery: add cros_update_cr50 action:
This action will determine the correct cr50 label for the DUT
inside the cros_repair plan. It can be either "CR50_PHASE_PREPVT",
"CR50_PHASE_PVT", or "UNSPECIFIED". And update the DUT's cr50 label
to the correponding one.

But since this action is just updating the cr50 label. It is allowed to fail
after recovery.

BUG=b:205156117
TEST=manual test

Change-Id: I65aa16b74db1b89294d432b9135b085e6ae2e62b