[recovery]: Verify if the servo device is connected to DUT with Type-A connection.
This commit is a part of the PARIS project.

Verify if the servo device attached to servo host is connected to DUT
with Type-A connection. This verification applies only to servo V4.

This commit also updates the check for Type-C connection, and uses
the pre-existing exec to fetch and verify servod control values.

Bug=b:198809511
TEST=build, manual test, unittest

Change-Id: Iae6c474260b48b563dcd5de297b68fef07bb176a