Remove shapiroc@chromium.org from go/src/infra/appengine/crosskylabadmin/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I361c5b3c80e17d21bfb9e7633cdc3954a069e084