Remove shapiroc@chromium.org from go/src/infra/libs/skylab/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ib340f0829392a34ef37bd508fa9bd4a905832a54