No longer look for python webinstall exe, and exclude 3.9.11.
These don't seem to be produced for releases starting with 3.9.11.
Look for the regular installer instead. We just use this as a check
that the release exists for Windows, for packaging we download the
MSI files directly.

Meanwhile, 3.9.11 is also missing launcher.msi for the amd64 release.
This looks like a temporary issue, so exclude 3.9.11 for now to keep
try jobs green. We'll revert this part once the missing file is
available.

No-Try: true
Change-Id: Idd211f6e21f56d3b73ecb5d581a31733ae898577