[cr-buildbucket] Only set the parent_task_id for the swarming task if the build doesn't have parent
Because of led builds, we have to keep the build.Infra.Swarming.ParentRunId longer, because the led builds don't have build id so their children cannot have parents.

But in the meantime, we want to make sure Buildbucket is the only service to track parent/child build relationship for real buildbucket builds.

This change only sets the parent_task_id for the swarming task if the build doesn't have a parent. In another word, only children of led builds have parent_task_id
set to their swarming tasks.

Bug: 1031205
Change-Id: I93f7e27a9456d7375b8ebb0f9509185a0f59bd60