recovery: Allow filter to fail when check for present files
When we read files from tar file we can fail if some files not found. Instead fail we can continue to check what is result we received from output and make decision based on it.
If any file is not. found then we still have to fail as we cannot perform fw update.

BUG=b:214315042
TEST=run local

Change-Id: I00752fe0f0251c5cfe34b2387908c776bec82775