[lkgr_finder] Replace failing Android builder
The `Android arm Builder (dbg)` is red for some time now and blocks the lkgr_finder from being updated. Given that V8 is the only consumer of the lkgr branch, we'd like to tailor the list of builders to match the config on our side.

This is the initial attempt to keep the lkgr branch updated. Further changes aim to only include tree-closers in this list to increase robustness of the finder.

Bug: 1321508
Change-Id: I1576b8066ec668559a07460d40fe81f46cb1e806
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3620275
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Auto-Submit: Alexander Schulze <alexschulze@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48671}