[recovery]: Updated the repair actions for ping verifier to make it same as legacy repair.
As a part of finalizing the changes to DUT repair config, we are
ensuring that the behavior of legacy repair is correctly replicated in
PARIS. This change in the config fixes a gap discovered in the PARIS
DUT repair config.

We need to ensure that all the repairs that are triggered by ping
verifier in the legacy repair, also get kicked-off in paris. Further,
they need to be executed in the same order as the legacy repair.

Bug=b:229786555
TEST=build, go test ./...

Change-Id: I0a3c7320066e32f07c12882866e8aa9144118dd4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3630851
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#48822}