[lkgr_finder] Increase allowed lag to 90 hours
In 1317669, we changed the lkgr finder recipe to fail for stale lkgr's instead of sending a notification only. This increased the failure rate for the lkgr builders (e.g. if a builder is not fixed on a weekend).

This proposes a change in the default configuration, although it only changes the chromium lkgr. The other lkgr_finder instances (v8, and webrtc) already override this configuration to a gap of 150 and a lag of 9.

Additionally, it adds the v8-infra team to the owners list since they started to own the lkgr_finder.

Bug: 1321508, 1317669
Change-Id: I331b9beee860da169650e5c79c1ef356d029760e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3629372
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48845}