[recovery]: Removed the provision repair as the recovery action for file system verifier.
As a part of finalizing the changes to DUT repair config, we are
ensuring that the behavior of legacy repair is correctly replicated in
PARIS. This change in the config fixes a gap discovered in the PARIS
DUT repair config.

In legacy repair, the verifier for file system is not a trigger for
the provision repair. This commit includes a change that causes
provision repair to not be triggered upon file system verifier
failure.

Bug=b:229786555
TEST=build, go test ./...

Change-Id: I0e1f8ee7330860def7f87cadad9392828b24bf16
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3632823
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#48852}