Update pytype wheel
Adds a version of the pytype wheel corresponding to the most recent
release (2022.5.5) and adds Mac as a supported platform since the
presence of prebuilt wheels for Mac implies it's supported.

This upgrade is being done because I'm running into what seems like a
bug in pytype with the currently available version (2021.11.2), so
hopefully upgrading to the most recent version will fix that.

Bug: 1315354
Change-Id: I506e9c522e55a2a20c4a4c1a6a06c077ec955df2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3639376
Reviewed-by: Brian Ryner <bryner@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#48915}