[labpack] Add Test SummaryMarkdown text
In preparation for adding a link to the logs, test to see
whether adding text to the summarymarkdown part of a step
causes us to render some markdown with a UUID in it.

If that's successful, we can replace it with formatted markdown
showing a link to various kinds of logs.

BUG=b:201240023
TEST=make test

Change-Id: I469034369639481faef929e8c19f14d55c61ef2b