[monorail] Update package.json.
Remove istanbul-instrumenter-loader as it is no longer maintained and
has no support for webpack 5. If we want test coverage we can integrate
a different package in the future.

Use react 17 instead of 18 because @material-ui/core@5.0.0-beta.5
requires 17. We can update MUI, but it's a larger migration that
will come in a separate CL.

Use lit-element 2 and lit-html 1, because I'm not sure if we're ready
for lit-element 3 and lit-html 2 yet.

Bug: monorail:10502
Change-Id: If7ed78fff3b22ba2ca98dc2d4b49df1df3eceb1e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3597717
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Andrew Chang <andrewjc@google.com>
Cr-Commit-Position: refs/heads/main@{#48945}