Update the priority of bootstrapped properties.
Currently, the bootstrapper gives higher precedence to properties
specified on the build than properties specified in the properties file.
The intention was that this would allow a properties to be overridden
when a build is scheduled. However, because certain properties are
duplicated into the properties file and into the buildbucket definition
so that they are available for querying, this prevents changes to those
properties from being tested on the CQ.

This change updates the behavior to give the properties in the
properties file higher precedence than those in the build's input, but
takes into consideration the requested properties. The requested
properties are the properties specified when the build is scheduled, so
giving those the higher precedence exactly fits the intended behavior.

Bug: 1324790
Change-Id: Ib20cd62639af19542bc4aefd898060af146046f7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3643276
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#48977}