[monorail] Add a newer version of Flask to requirements.py2.txt.
Follow-up from https://crrev.com/c/3643477
We're seeing the following exception:
    from werkzeug.middleware import dispatcher
  ImportError: No module named middleware

This is because werkzeug==0.11.10 bundled with App Engine Python 2 [1]
is too old and doesn't contain the middleware module.

In order to work around this, we add Flask and Werkzeug to monorail/lib/
using requirements.py2.txt instead. That gives us the freedom to choose
whichever Flask version we like.

The caveat is that we can only add pure Python dependencies this way,
not any packages with binary wheels. One of Flask's dependencies,
MarkupSafe, is only available as a binary wheel. Thankfully, App Engine
has MarkupSafe==0.23 available as a built-in library. So I've added
MarkupSafe to app.yaml (for production) and requirements.dev.txt (for
local development).

I've chosen Flask==1.1.4, which is the last version of Flask that
supports Python 2, and luckily it's still compatible with
MarkupSafe==0.23.

[1] https://cloud.google.com/appengine/docs/standard/python/tools/built-in-libraries-27


Bug: monorail:6511
Change-Id: I9f4c9409f03d8e9ded5baebfc9d9a7fee45103ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3645664
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48996}