Add a polymorphic flag to the chromium bootstrapper.
This will be used to modify the the bootstrapper's behavior for builders
that will be bootstrapped with other builders' properties files in order
to perform operations using the same settings as those other builders.
Currently the flag does not change any behavior, but it needs to exist
so that the findit-rerun builder which is polymorphic can set it before
the default behavior is changed to prefer the builder properties over
the build properties.

Once the default behavior is changed, the polymorphic flag will cause
the build properties to be preferred. The expectation is that any
properties that the polymorphic builder is actually setting in its
definition are required for proper operation of the builder (e.g.
recipe).

Bug: 1324790
Change-Id: Ia8db247439eef97dff0986b19256bed238e0eab3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3646497
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#49015}