[monorail] Remove karma-parallel.
karma-parallel has a dependency on a high severity vulnerability. The
package is no longer maintained and hasn't released in 3 years.

The simplest option to resolve the vulnerability is to remove the
package. I estimate that parallelization saves us ~3-6s on unit test
runs. So it's not a huge difference.

karma's documentation says it has the `concurrency` option that is
enabled, but I haven't managed to make it speed up the test runs
to the same extent as `karma-parallel`.

BONUS: Run `npm audit --fix` on infra/appengine/.

Bug: monorail:10502
Change-Id: I98fb4530bd53b9ef451ea7cfb265924d3fab53b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3650461
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Auto-Submit: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49032}