[QEMU] Add qemu recipe module
Bug: b/232831070

This is a recipe_module for downloading and utilizing qemu binaries on
our linux bots. The binaries are currently hosted in a experimental
location. They will be moved a proper location once we can dial in the
configuration flags required to build the binaries for online builders.

Context: go/win-online-design

Change-Id: I7d67644d4988c2d5db1b785f7a08962f00ba305a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3646487
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49168}