Refactor gerrit operations in chromium bootstrapper.
The bootstrapper occasionally fails in the presence of gerrit
replication lag because the RPC to get the change info will succeed but
it won't contain the patchset corresponding to the GerritChange present
in the build. To address this, instead of the gerrit client having
operations for getting the target ref of the change and the revision of
the change and the patchset, the client interface has been updated with
a single operation that will get the change info and check that it
contains the patchset. In the event that the change info does not
contain the information for the desired patch set, it returns an error
with codes.NotFound so that the operation will be retried.

Change-Id: Ifa00bdbbac323c8c9f6e21fc710a50345a6f1799
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3656452
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#49216}