Move swarm image builders to push images to chops-public-images-prod
Currently, the bots that use these images download them via old
service account JSONs (deployed via puppet).

b/233998033 is yelling at us to get rid of old static service account
keys. So the easiest way to get rid of the keys for these bots is to
make their images world-readable so the bots don't need to auth. This
should be fine since there's nothing internal/confidential in the
images. (The images can be built identically by anyone.)

Their prev project "chromium-container-registry" is not configured for
its images to be world-readable. We could flip the toggle to make it
so, but there are other project's images in that registry:
http://shortn/_oDcQCjge6K

It's unclear those other images are safe to make public. So instead
this moves the docker images to the "chops-public-images-prod"
registry which is already configured to be public.

This should let the bots download the images without the use of a
static service account key.

Bug: b/233998033
Change-Id: Ib54bd799641a858fe9c78c42495cacf07d76e826
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3670048
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49348}