infra: Replace the infra-continuous-bionic-arm64 bot with a focal one
This moves the following bot to focal:
https://luci-milo.appspot.com/p/infra/builders/ci/infra-continuous-bionic-arm64

Ran this on led and the build result appears unchanged:
https://ci.chromium.org/swarming/task/5b1b680348cf8b10?server=chromium-swarm.appspot.com

This will free up its machine:
https://chromium-swarm.appspot.com/bot?id=build41-b4

That swarming bot is unique in that it's:
- ubuntu
- arm
- not inside docker

All three of those things are needed to build docker images that
run on arm machines. (Last I checked you can't cross compile images
nor build inside another container.)

And as part of b/233998033, we're moving the swarm-arm64 docker images
to a new registry. The new registry has write access restricted in the
infra CI pool:
https://ci.chromium.org/ui/p/infra/builders/ci/swarm-docker-arm64-image-builder/19/overview

So rather than giving that pool write access, we'll instead move that
builder to infra-internal.prod pool. Which means this machine
(build41-b4) will need to move to an internal vlan. Which means it
won't be able to run infra-continuous-bionic-arm64's builds any more.
Which means infra-continuous-bionic-arm64 will break unless we move it
to an arm pool we have extra capacity in. And we have some focal arm
bots:
https://chromium-swarm.appspot.com/botlist?c=id&c=task&c=os&c=status&d=asc&f=cpu%3Aarm64&f=pool%3Aluci.flex.ci&f=os%3AUbuntu-20.04&k=os&s=id

Hence the s/bionic/focal/ on this bot.

Bug: b/233998033
Change-Id: Ie30a72fe5271017c53efac1f1e3f4b2ab1dc2a53
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3671758
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49356}