[Weetbix] Limit concurrent requests in ingestTestResults.
GAE instances are occasionally shutting down due to excess memory
usage. Cloud logging indicates this happens when there is a mass
pileup of result-ingestion requests on the same instance.

The root cause appears to be the significant asymmetry in request
sizes, which is not known the GAE when scheduling requests on
instances.

Work around the issue without capping overall concurrent requests
excessively by using a semaphore. Limit concurrent requests per
instance to 20 as current instances are doing nowhere near 50
anyway and this will limit the size of any pileup.

BUG=1314194
TEST=Integration tests

Change-Id: I39f93777a4200c0171a3f24049794b1c6ff518fe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3671056
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#49362}