[Weetbix] Pass TaskIndex to clustering ingestion.
ChunkIDs are generated based on the Ingested Invocation, Task Index
(i.e. QueryTestVariants page number) and chunk within that page.

The task index was not being passed, leading to duplicate chunk
IDs. Because there was functioning concurrency control on chunks,
this should not have led to data inconsistencies (e.g. chunks in
GCS not matching chunks in Spanner or failures in BigQuery) but
it did lead to transient "ingesting for clustering: concurrent
modification to cluster" failures in the ingestion task queue,
and some failures being dropped in ingestion (as ingestion for a
chunk is skipped if it already exists).

This would only have affected builds with more than 10,000 failing
test variants. The issue was caught in dev so there should be
no prod impact.

BUG=1314194
TEST=integration tests

Change-Id: I2f1a067122abe377d5de37cc03049bc799e5f789
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3672989
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#49377}