Update test_plan_generator README
This CL updates the test_plan_generator with infra/infra-specific
instructions. The new docs still leave much to be desired -- the
fact that sub proto contents are serialized makes it impossible
to directly edit the input JSON. In a follow on CL we'll want to add
a workflow for generating arbitrary input JSONs.

BUG=b:234019083
TEST=followed new documentation

Change-Id: I24c9bc8fdadfdf90e337780890b94a91227da631
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3681919
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Benjamin Shai <bshai@google.com>
Cr-Commit-Position: refs/heads/main@{#49419}