Remove all login-flow code from swarm-docker services
We're switching to using world-readable images so we can get rid of
the static service accounts. Which means all this login code won't be needed.

We could keep the code checked-in but unused in case we ever try
using internal images. But that could be YAGNI, so this just tears all
the code out. Git history will remember it if it's ever needed again.

Bug: b/233998033
Change-Id: I11529e002bd2f4f66131197ff6b1a0ed99c04ada
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3682078
Reviewed-by: Elliott Friedman <friedman@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49420}