[lkgr_finder] Update local repository and retry for invalid revisions
On a new bot, the lkgr finder [1] needs to checkout its underlying repository. This takes around 60 minutes. After the checkout, the finder determines the new lkgr branch and looks up unknown commit positions in the previously checked out repository.

In some cases, the checkout is older than the successful revision determined from the builders, and the newer revision cannot be mapped to a position [2].

In those cases, we will fetch the repository again and retry to map the revision to a number.

Additionally, the OWNERS file is moved one directory up, since the V8 infra team started to own the lkgr_finder service.

[1] https://ci.chromium.org/p/infra/builders/cron/chromium-lkgr-finder
[2] https://logs.chromium.org/logs/infra/buildbucket/cr-buildbucket/8812795109746768129/+/u/calculate_chromium_lkgr/stdout

Bug: 1330292
Change-Id: I7d6df9b88af3c5bbbba638e7a4f2a04d33d7df55
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3677252
Commit-Queue: Alexander Schulze <alexschulze@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49473}