Add -vpython-tool for vpython_ng
Tool mode is implemented as subcommands in vpython to support install,
verify and delete. To simplify it in the new implementation, we change
it to be implemented as a normal argument.

Delete command is not implemented because we don't have a clear use case
for now. Verify is using cipd cli instead of cipd's go library.

Change-Id: I6e1f65859e109c3eda7be3d26c928644019912b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3682926
Reviewed-by: Brian Ryner <bryner@google.com>
Auto-Submit: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49506}