Add psmisc to docker swarm bots' images
It appears that rolling them forward as part of b/233998033 caused them
to lose the `killall` binary in the image. This is called in a few
places:

https://chrome-internal.googlesource.com/infradata/config/+/43d061a59b3c716665988f7f55f70f1c5153a6df/configs/chromium-swarm/scripts/bot_config.py#318

https://source.chromium.org/chromium/infra/infra/+/main:luci/appengine/third_party/python-adb/adb/contrib/adb_commands_safe.py;drc=a126ae6e55bdb50dba73c3b2ff43ae4140a10968;l=71

So when we rolled the image forward, those callsites ended up failing.
And that 2nd callsite is needed by the bot to clear a test's lingering
ADB proc from holding onto the device. Since the bot never ends up
killing ADB, it can't talk to the device and assumes it's a standard
linux bot. Hence why we saw android bots taking linux tests in
crbug.com/1333559.

Bug: 1333559
Change-Id: I4a113e12e47e277cbf8719d207f1c93e5d68e11e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3692175
Reviewed-by: Elliott Friedman <friedman@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49560}